Remove boolean type schema condition for `essentialRequirementsMet`
We remove this as `essentialRequirementsMet` must still be of enum
'true' and therefore the boolean type condition was not adding any
extra value. However, because we had these two conditions, if the
API encountered a string both conditions would fail. In python 3
the API would then return a random error from the above two
conditions due to randomised ordering of hash tables. Now, only
having one condition that can be broken we are certain what our
error message would be.