Add method app/models.py::Framework::get_supplier_ids_for_completed_service
This adds a method on Framework to get all supplier id's who have managed to successfully complete inputting a service.
This info is added to the export_users_for_framework view in views/users.py.

It also adds a new py.test fixture in conftest.py which supplies a draft service.