Validate length of essentials for new responses
Previously we were only validating the length of the
essentialRequirements key for old brief responses. This was because the
new flow didn't use this key.

Now we are collecting evidence in a list for the new flow we're using
this key again and so can use the same validation technique.