Add test case for invalid brief response schema validation
We add a test case for when evidence is provided but `yesNo` has
been answered False. We don't expect to have to validate data
like this as evidence should be stripped if you have answered
False but we make sure that if somehow evidence did get provided
that our schema doesn't accept it.