Add validation to handle boolean instead of dicts in dynamic lists
Legacy data for nice to have requirements is sent as a list of
booleans. New data will be a list of dicts. We should therefore
validate correctly if you tried to pass legacy data to our new
JSON schema.

As we don't expect to receive booleans as part of a dynamic list
from the frontend, we don't need to concern ourselves with
having a "pretty" indexed error and instead just need to pass
something readable e.g. "True is not of type object".