Remove import_service route + nose
This commit is a bit complicated because everything is done
together, so here's a list of what it contains:

- removed `import_service` route
- refactored test_drafts.py
  - using `create_dummy_service` instead of the PUT route
    to create services in test setup
  - creating a SaaS service insteaad of IaaS
    (doesn't seem to make much difference)
  - wrote a class method to update a service
  - switched all nose test stuff for pytest equivalents

The last one is interesting.
I found a [repo](https://github.com/pytest-dev/nose2pytest) that
will convert nose methods into py.test asserts, so I've run it
in this file and it looks like it works pretty well.

Have separated out this last part in subsequent commits as it
seems like quite a large change.