Extend setup_dummy_service() method
`setup_dummy_service()` would previously accept a handful of
arguments when creating a service, but ultimately it was no match
for the PUT (or even POST) route, which is what we started using
in our tests. (ie load a prebaked fixture and import it using PUT)
Extended the setup method to take kwargs, which we pop useless things
out of and then feed the rest directly into our new Service.data.
This allows us to dump the entire contents of a fixture into
`setup_dummy_supplier()` and have it create a complete service for
us, something that was previously only possible using the API
directly.