Update API user creation error message
Small change.
Previously, sending in user creation data where role='supplier'
but there was no supplier id would retun an error message asking
for a `supplier_id`. This is the name of the database column, but
the JSON schema validates on `supplierId`, so updating this message
seemed like the obvious thing to do.