Add final commit to supplier/brief setup
Most of the methods in these fixtures take care of their own
commits, except for services and briefs, where I suppose the
idea is that we can bulk-add them.

Added a commit at the end of ever helper file (except the bulk-add
ones) instead.

Thought it might make the tests noticably longer to run, but it
seems as though they take about the same amount of time.