Create migration for draft dos1 briefs to dos2
When DOS1 is closed for brief submissions and DOS2 opened, any supplier
who has a draft DOS1 brief will effectively lose it, as they'll be
unable to publish it.

This migration will move all draft DOS1 briefs to be draft DOS2 briefs,
allowing them to be published. Which framework version the brief
goes into doesn't affect buyers - they are eligible by default.

This will need to be run immediately after DOS1 is closed so that no
buyer attempt to publish their DOS1 brief in the period between DOS1
closing and the migration happening.