Make migration agnostic of env framework id's
Preview is using a different id for dos2 than staging and production.
This means the id's for the frameworks can't be hardcoded.

This change checks what the id's are before changing anything which
means it'll work in any environment.

This could probably have just been done for the DOS2 id as that's the
only one that changes between environments, but better safe than sorry.