Update brief responses schemas with new followup rules
Changes to the followup schema generation preserve the existing
behaviour but move some of the rules around to work when used in
the API.

Eg dropping the lead-in question.id from the oneOf schema required
lists allows us to use page_questions and partial service updates,
which raised validation errors otherwise. Instead, it moves the
lead-in question to the parent schema's required list, since if it's
not optional then it's required in both cases.

Another change is dropping the property description for the followup
question from the oneOf subschema. The property is described in the
parent schema, which has access to the question type to generate a
correct one, so the only thing we need to add is to mark it as
required.