Append "unexpected followup" errors to _form errors
Trying to display "unexpected followup" error under a related
field key relies on the validation error order which appears to
be different in python2 and python3.

Since the error isn't really limited to one of the questions, but
rather describes an issue with the combination of multiple field
answers it makes sense to add it to the `_form` errors list instead.