Explicitly name briefs framework_lots foreign keys
Due to the new naming convention which now matches the original pattern,
when this migration was applied to an empty database it tried to create
two foreign key constraints with the same name.

Before, when the naming was left to SQLAlchemy, the name was
auto-incremented to be prepended with a '1' to prevent a clash. Now that
a naming convention is defined this doesn't happen so an error is
thrown.

Explicitly naming this constraint stops this clash and also keeps the
name in sync with the prod database, even when applyng the migration to
a clean database.