Remove plain text logs from live
Version 25.2.0 of the utils will by default only write logs in json
format. If the `DM_PLAIN_TEXT_LOGS` env variable is set True, then the
logs are plain text. This is being set in dev and test to make
developers jobs easier.

The other things that 25.2.0 does is write logs to stdout, unless
`DM_LOG_PATH` is set. This is useful for the PaaS, but also useful for
dev and test environments again. To be clear, we're only using this
variable in the ElasticBeanstalk environments where we need to write
logs to files.