Add brief to brief response serialisation
This commit adds the related brief to the serialisation of a brief response. The brief response is not very useful without this and we would have had to perform n+1 api requests whenever fetching the brief response.
In assessing the SQL for this change an n+1 SQL query was discovevered on briefs - clarification questions. This was remedied by adding the lazy='joined' kwarg to the relation.
As a sub object the brief serialisation on the brief response ha been kept minimal to allow for refactoring at a later date. To this end just 4 keys were added: id, status, title and applicationsClosedAt.

Tests were updated