Pull in v 27.1.1 of utils and 8.10.2 of apiclient
The new utils version has a fix for header's needing to be strings in
python3. The two major changes, 26 and 27 don't affect the api.

The new apiclient version has a fix to only require enum34 if the python
version being used is 3.4 or below. The major version bump, 8, doesn't
affect the api.