Add columns to suppliers table
The following data is moving from framework-level supplier declarations to the
supplier account level:

* Organisation size
* VAT number
* Non-companies-house registration number (for non-UK companies)
* Registered organisation name (as opposed to marketplace supplier account name)
* Registration date
* Trading status
* Registration country

This adds the relevant columns to the supplier table. All are nullable string fields, except
registration date which is a string in the declaration data we have, but we are moving to
a proper DateTime field at the supplier account level.