Allow supplier updates with no companies house number
Companies house number is optional but the validator broke with `TypeError: expected string or bytes-like object` if the value passed in is None.  This allows validation to pass for empty companies house number, and adds a test to guard against regression (no existing tests failed when the validator was added because the test supplier had companies house number set).