Return projectId in list_project_services services
The services being returned by the `list_project_services` endpoint did
not include the project id associated with the request.

This made dealing with the data returned difficult, especially if making
multiple calls for different projects, such as when extracting data to
analyse.

This also updates the test of the response to assert that the data is
correct, and not just that the keys are in place.