Set SQLALCHEMY_TRACK_MODIFICATIONS = False
We currently get 1061 warnings when we run our tests, as well as a warning every time we start the API, about how "SQLALCHEMY_TRACK_MODIFICATIONS adds significant overhead and will be disabled by default in the future".

These warnings are annoying. Making this change reduces the number of warnings to 16. (The remaining 16 are to do with yield tests being deprecated and I'm not going to fix that right now.)

I have checked and we're not using the feature in our code anywhere, so seems like turning it off is the best thing to do (see the accepted answer here for details:
https://stackoverflow.com/questions/33738467/how-do-i-know-if-i-can-disable-sqlalchemy-track-modifications/33790196#33790196)