Standardise serialized json responses
We very often return JSON responses which are
- single serialized objects from our models
- lists of serialized objects from our models
- paginated lists of serialized objects from our models

However, when doing this in the past we have not had consistency.
For example, some lists of objects may include meta such as the
total number of results. Other responses may include links to
their own resources.

This introduces the wrappers for those common types of responses
so we can introduce consistency. This means all lists of results
will contain meta totals for example. If we want to add links to
all our results then that can easily be done in one change.

To implement this around the codebase there were a few exceptions/
things that needed changing. Some times we are not serializing
objects but instead jsonifying dictionarys of data - this case has
been ignored. A handful of our list responses were including either
an empty links dict or a dict containing a link to the resource it
self. These have been removed for consistency as it was rarely done
and our front end does not use them. We can choose to implement
this for all responses if we so choose at a later point.

This commit also tidies up a few remaining areas:
- fixing an incorrect status code from 200 to 201
- moving the number of paginated results for buyer domains into
the config
- ensuring status codes are explicit on all return values from
our views
- imports adhere to pep8 standard (https://www.python.org/dev/peps/pep-0008/#imports)
- make the resource name more DRY e.g. not writing "briefs"
for the key of the response for every briefs endpoint