Add comments as to why no indexing of draft briefs
We don't need draft briefs in search results so we don't index them on
creation or update.

There is an argument that we should be syncing our elasticsearch index
with our db for everything by including an ORM hook at a lower level,
however that's something we might look at in future.

For now, drafts will get synced by the overnight batch job. If they are
published during the day the entire document gets reindexed, so all is
well.