Upgrade pytest-cov to remove deprecation warning
pytest was giving a deprecation warning about `pytest_funcarg__` prefix
being deprecated in pytest 4.0, in favour of just using the fixture
decorator.

The version of pytest-cov we were using was still using it, but was
fixed in version 2.3.1. Upgrading removes the warning.

Jinja is also updated, as 2.10 has been released and all the packages we
have that require it specify >= 2.4. There are no breaking changes.