Bump api client version to 14.0.0
It allows `make_iter_method` accept more than one model name. We need
this so we can upgrade the search-api to return `documents` instead of
just `services`.

The direct award `lock_project` view uses the search-api to fetch
service ids. We need to bump the api-client so this view doesn't break
when the search-api starts returning documents.