Allow IntegrityErrors to be serialized
We had seen 500s in the app if an IntegrityError, e, was caught and
we passed `e.orig` into the `abort` function. This is because
`e.orig` refers to the original DB-API IntegrityError that is
raised by the database adapater (psycopg2 in our case). The pyscopg2
IntegrityError is not JSON serializable and we ran into problems.

We can instead use `format` to allow our exception to be
displayed as a string. This is a pattern we are already using in
the many other catching of IntegrityErrors across the app.

http://docs.sqlalchemy.org/en/latest/core/exceptions.html#sqlalchemy.exc.DBAPIError