Add validation for country-picker values
The accessible country picker select widget we're now using stores vales as "country" or "territory" followed by a colon followed by a two-capital-letters country code.

We want to enforce this format in the database, for consistency.

Adding the validation here now will enforce it for all future updates. A migration will be needed to update the existing "gb" values to "country:GB".