Fall back to created_at if no _logged_in_at timestamp
When a user is created they are automatically authenticated when they first set a password.  This means that the "lastLoggedIn" timestamp is null, even if they have done work (e.g. created a brief) that you would think needs a login.  This is confusing for admin users who see users listed who have apparently never logged in.

By falling back to the created_at timestamp we accurately reflect that they were logged in when they first created the account, even if they have never been through the normal log-in flow.