Add company details complete column
 ## Summary
Add a new column to the Suppliers table that tracks whether or not a
supplier has confirmed that the company details that have been submitted
are complete and correct. Defaults to False in the migration so that all
existing suppliers will get a one-off chance to update their existing
details. After this, the frontend will insist that suppliers go through
support to get further corrections (but the API will remain editable
regardless of the flag's value).

 ## Ticket
https://trello.com/c/7UIX3xPs/59