Allow clearing other reg number
We want to allow suppliers to update some of their company details until
they click a big green 'confirm details are correct' button. This means
that a supplier might provide an 'other company registration number'
incorrectly and then want to change it to a Companies House number,
meaning we need to let them post an update which clears out their other
company registration number.

They can also provide a VAT number and then decide they shouldn't have,
but when a supplier declares they haven't registered for VAT we store
the value "Not VAT registered" in the column rather than leaving it
blank, so we don't need to worry about allowing null updates there.