Include frameworkSlug in framework update audit event
Without it there was no way to quickly determine which framework an
audit event was for.

This also adds a db session flush after creating a framework. This is
because without the framework has no id when passed to the
create_framework audit event, and the object_id gets stored as `None`.