Remove serializeable isolation level from draft services
We noticed that when under high load, we were seeing some transaction
errors from psycopg when updating draft briefs:
```
sqlalchemy.exc.OperationalError:
(psycopg2.extensions.TransactionRollbackError) could not serialize
access due to read/write dependencies among transactions
DETAIL:  Reason code: Canceled on identification as a pivot, during
commit attempt.
HINT:  The transaction might succeed if retried.
```

We think this was due to specifying that the transactions were fully
serializeable, which put a lock on more resources than necessary. By
updating the query to use `with_for_update` we can make sure only the
row we're interested in is locked.

We also noticed that the query for the draft service was joined, and so
each time we fetched the draft service from the db we were pulling back
_loads_ of stuff we didn't need. Frameworks, lots etc. Have chance to
lazy loading so we drastically reduce the cost of the query.