Update validation schemas
 ## Summary
Updates the API validation for G-Cloud 10 based on recent changes to
content and some logical changes to security questions. This commit
includes some changes to the ordering of objects in the G9 schemas.
Using DeepDiff (https://github.com/seperman/deepdiff) with
ignore_order=True, we see that the old and new schemas have identical
content - it's just ordered differently, so I include those changes in
order to bring them into the fold. Otherwise, whenever we run the
generate-schemas script again, we'll keep having to ignore and drop
these updates.