Return only a count of created drafts when copying all
When a supplier has a large number of services which are all copied to
drafts in one go, the response was quite large and slow. The highest
number of services we've seen in one lot is 280. Serializing them all
and sending them in the responses is unnecessary. All the supplier
frontend was doing was counting the number of serialized drafts, so we
can just send that instead.