Organisation size from supplier
 ## Summary
A supplier's organisation size used to be provided during the framework
application process as a question in the declaration. It is now provided
at the supplier account level and is a feature of a Supplier record.
Therefore, when returning the organisation size in a brief responses, we
no longer need a join to the SupplierFramework table and can instead
pull it directly from the Supplier.

 ## Ticket
https://trello.com/c/PNPqPL8b/66-brief-responses-still-look-for-supplierorganisationsize-in-the-declaration-rather-than-the-supplier-table