Store framework dates in DB
 ## Summary
At the moment we have strings describing dates for important events in
the framework lifecycle stored in the dm-frameworks repository in a YAML
file. We also duplicate one of these dates on the framework records in
the API (frameworkCloseDate). This leads to some duplication and the
possibility for these to get out of sync, or for one or the other to not
get populated (which indeed happened recently for G-Cloud 9).

We will store the dates in UTC, so changing it to a local timezone
format will be a job for the frontends. This feels like the safest
approach but does mean we need to be considerate around setting dates
during British Summer Time as we'll need to remove an hour from the
time that we're want to save. I've named the new fields with a `UTC`
suffix to make this apparent, and temporarily duplicated the
`applicationCloseDate` field under `applicationCloseAtUTC` to be
consistent.

By storing these dates solely on the API, we can return to a single
source of truth and by storing them as datetimestamps rather than
strings, we can easily convert them into whatever format we need on the
frontend. It also leads us towards being able to submit these days e.g.
via the admin frontend, leading to management of more framework
configuration by non-devs.

We will remove `dates.yml` files eventually, as the fields added here
replace all of the dates we store. See:
https://github.com/alphagov/digitalmarketplace-frameworks/blob/e1f8e5c4a1a98d817e1bc9ff90ace85c1a0762c6/frameworks/g-cloud-10/messages/dates.yml

Follow-up PRs intend to make these non-nullable fields. We can enter
provisional dates before we pin down a timetable, as we currently do in
the frameworks repository, but as things in the frontend depend on dates
being available, it feels far safer to require that *some* date is
provided rather than not.