Revert migration 1160
 ## Summary
Migration 1160 set some framework date columns to non-nullable, but more
importantly removed the `application_close_date` column. This column was
not removed from the Framework model, so when the migration went in,
SQLAlchemy refused to work with the database as its idea of a Framework
didn't match the database's.

By adding the column back in temporarily, we can correctly remove the
column from the model and ship another migration to re-apply 1170
afterwards.