Apply migration 1160 again
 ## Summary
Migration 1160 was shipped in the same release as the code which removed
the column from the model, which breaks the app in the interim period.

Migration 1170, in a separate release, has added the column back to the
database and doesn't have the column in the SQLAlchemy model. When that
successfully goes out, this PR brings in migration 1180 that removes the
column from the database altogether.