Add audit event for failed login
 ## Summary
We have an audit event type defined for `user_auth_failed` but this
doesn't seem to have been added to the appropriate API endpoint. We've
had a support query come in that we are unable to properly investigate
due to being unable to tie up login attempts with requests. This now
adds an audit event on failed login which will store the email address
of the account, the client IP making the request, the hashed password
and the number of failed login attempts currently on the account (after
adding one for the current requests).

 ## Ticket
https://trello.com/c/iSAzWgQk/466