Add framework components to model
 ## Summary
We currently enable certain pages/journeys based on the name of the
framework (e.g. catalogue search for 'g-cloud' and competition for
'digital-outcomes-and-specialists'). This doesn't scale well and, based
on our current understanding, isn't really reflective of how frameworks
are composed.

This adds two booleans to the framework table: `has_direct_award` and
`has_further_competition`. We will set `has_direct_award` to True for
G-Cloud frameworks and `has_further_competition` to True for Digital
Outcomes and Specialists frameworks. We can then gate journeys based on
the values of these booleans. Based on our current understanding, a
framework could theoretically have both elements, allowing both
journeys to be used.

The migration for this initial PR will make the fields non-nullable and
give them a default of False. We will POST to the API once this is in to
assign the correct states. After that is done, a further migration will
add a constraint enforcing that at least one of the two components is
present on a framework record.

 ## Ticket
https://trello.com/c/WruHrcBf/456