Enforce framework component constraint
 ## Summary
Enforces a constraint of the Framework model to ensure that each
framework record has at least one of the two components: direct award or
further competition. In theory a framework can have both, but the
Marketplace only supports these two frameworks so it must have at least
one of them or we will have no features to enable for it.

 ## Ticket
https://trello.com/c/WruHrcBf/456