Preparing to remove `FrameworkFramework`
 ## Summary
We want to deprecate the variable name `frameworkFramework` and replace
it with `frameworkFamily`, which is easier to understand. We need the
API to remain compatible with the frontends,  which will still expect to
find `frameworkFramework`, until we get around to updating them. So for
now, we will just add the new key `frameworkFamily` and once this goes
out we can start to transition the frontends over to using the name.

 ## Ticket
https://trello.com/c/iqFcVKpd/14