Update supplier framework details on confirmation
 ## Summary
A supplier has to confirm that their company details are correct as part
of an application to a framework. They need to do this every time they
apply to a framework to ensure that they are still the same company. To
track this, we need to have two concepts of company details being
confirmed:

1) Company details get confirmed at the account level. This confirms
that their account is configured correctly to represent the entity they
will be applying with (specifically this locks down four attributes for
now: company name and number, vat number, and duns number).
2) Company details are confirmed as being accurate for a specific
framework application. They must confirm the details are still correct
for the application they are making, as some of these things change over
time. If a significant change happens (e.g. new duns number), they will
need to apply with a new supplier account.

An application is made when the details are confirmed for that framework
application, so we will update the logic for these checks as well.

For G-Cloud 10, we injected all supplier company details into the
SupplierFramework when the Framework was set to pending. This is no
longer required now that we are injecting the details on every relevant
update, so we'll remove this functionality.

 ## Ticket
https://trello.com/c/wrz3NRG3/171