Centralise application status
 ## Summary
Total strawman for centralising the logic around whether an application
has been made to a framework. This logic is currently fractured within
the API and also duplicated on the supplier-frontend. It's a prime
candidate for bringing together to one place. This feels really
important given how core the idea of making, or not making, an
application is to the platform.

The two endpoints in the API that use this piece of information are
`export_suppliers_for_framework` and `export_users_for_framework`. Part
of the reason for this logic not being on the SupplierFramework before
is probably to keep these endpoints performant by doing one of the
checks (whether a supplier has submitted a service for the framework)
performant by running it only once, rather than checking for each
individual application. This is a valid concern and I'm open to ideas
as to how we can keep this endpoint returning in an acceptable timespan
while also keeping the application_status logic in a single place.

One of the tests here breaks because it submits a DOS application and
asserts that no application has been made, expecting it to fail on the
fact that company details must be confirmed. This is a slight misnomer,
as company details didn't need confirming for DOS1. The code is behaving
correctly for the framework under test; the test is currently invalid.
Fixing it may also require a larger refactor. There's also the question
of whether we need to keep `application_status` valid for historical
frameworks. Ideally I'd like that to be the case, but it's probably
unlikely that we'll actually export users for old frameworks anyway?