Return contact information consistently
 ## Summary
If we don't have an order on the Contact Information, they will be
returned in an undefined order each time. We only technically support
multiple entries at the moment, so we won't have encountered any issues
with this in production as, in reality, all suppliers only have a single
contact information entry. Our tests, however, have two contact
information entries for the supplier, so this has introduced a lack of
determinism in some cases.